fix(develop): 修复了由于从Redis中获取数字默认转换为Integer导致的读取错误问题

+++ personal-blog-common/src/main/java/com/gewuyou/blog/common/dto/ArticleAdminDTO.java
@@ -49,7 +49,7 @@ public class ArticleAdminDTO {
      * 访问量
      */
     @Schema(description = "访问量")
-    private Integer viewsCount;
+    private Long viewsCount;
 
     /**
      * 分类名称
